diploma projects management app v3.0 Model Layer Complete.Added Strategy for Best Applicant. Service Layer Complete(exception to ProfessorServiceImpl with 1 todo).Revised and Complete DAO Added Student/Professor Main menu. Sing in / Sing Up revised. Professor Controller / Student Controller ready. Professors can use the app already.
Fixed Bugs:
Now properly using current average Grade as Double instead of Int.
Properly Redirecting to Login After Signing up and setting a profile.

+++ src/main/java/diplomas_mgt_app/model/strategies/BestAvgGradeStrategy.java
@@ -10,7 +10,7 @@ public class BestAvgGradeStrategy implements BestApplicantStrategy{
     }
 
     public int compareApplications(Application app1, Application app2) {
-        return Integer.compare(app2.getStudent().getCurrentaveragegrade(), app1.getStudent().getCurrentaveragegrade());
+        return Double.compare(app2.getStudent().getCurrentaveragegrade(), app1.getStudent().getCurrentaveragegrade());
     }
 
     @Override
