+++ src/main/resources/DDL/secure_users_directory/students.sql
@@ -6,7 +6,7 @@ create table student
     last_name                                  varchar(45) null,
     email                                      varchar(45) null,
     year_of_studies                            int         null,
-    current_average_grade                      int         null,
+    current_average_grade                      double        null,
     number_of_remaining_courses_for_graduation int         null,
     username text not null
 )
